Fix test failures caused by incompatible is-generator-function version
Pin is-generator-function to 1.0.10 to resolve compatibility issues with tsx loader in mocha tests. Version 1.1.2 introduced breaking changes in how generator-function module is loaded, causing "getGeneratorFunction is not a function" errors when Koa attempts to validate middleware handlers.

Related issues:
- https://github.com/inspect-js/is-generator-function/issues/46
- https://github.com/ljharb/async-function/issues/4

This is a temporary workaround until the upstream libraries fix compatibility with tsx/CommonJS module loading.